backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
mysqlclient>=2.2
python-dotenv>=1.0
